VThreadID: widen to uint64
Convert VThreadID to uint64.

+++ open-vm-tools/lib/include/vthreadBase.h
@@ -83,7 +83,7 @@ extern "C" {
  * Types
  */
 
-typedef unsigned VThreadID;
+typedef uint64 VThreadID;
 
 #define VTHREAD_INVALID_ID    (VThreadID)(0)
